requirements.txt
pytorch==1.8.0
cudatoolkit==10.2
tensorboard==1.15.0
matplotlib
pandas
rich
pyyaml